Brings together two lines that had diverged since 1.22.0: the local broadcasting
work and releases 1.23.0-1.25.0 from origin (Telegram reactions and polls, rich
message support, Sonnet 5 and adaptive thinking).

Conflicts resolved:
- build.gradle.kts: keep Kotlin 2.3.21 and java-telegram-bot-api 10.1.0 from
  origin, keep the removal of the unused github-api dependency, version 1.26.0
  since the merge adds broadcasting on top of 1.25.0
- CommandUtils.kt: keep origin's wording of the doc comment

+++ src/main/kotlin/space/davids_digital/kiri/integration/telegram/TelegramMessageBuilder.kt
@@ -9,7 +9,7 @@ class TelegramMessageBuilder {
     val attachments: MutableList<TelegramOutgoingAttachment> = mutableListOf()
     var replyMarkup: TelegramInlineKeyboardMarkup? = null
     var disableNotification: Boolean = false
-    var messageThreadId: Int? = null
+    var messageThreadId: Long? = null
     var replyToMessageId: Int? = null
 
     fun html(html: String) {
